# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ratchaburi Community Jobs API",
    "version": "1.0.0",
    "description": "API documentation for Ratchaburi Community Jobs platform - A comprehensive job matching platform for local services in Ratchaburi province",
    "contact": {
      "name": "Ratchaburi Community Team",
      "email": "support@ratchaburicommunity.co.th"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://ratchaburi-api.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Provider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Provider ID"
          },
          "name": {
            "type": "string",
            "description": "Provider name"
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "line_id": {
            "type": "string",
            "description": "LINE ID"
          },
          "service_category_id": {
            "type": "integer",
            "description": "Service category ID"
          },
          "location": {
            "type": "string",
            "description": "Address"
          },
          "district": {
            "type": "string",
            "description": "District"
          },
          "subdistrict": {
            "type": "string",
            "description": "Subdistrict"
          },
          "province": {
            "type": "string",
            "description": "Province"
          },
          "description": {
            "type": "string",
            "description": "Service description"
          },
          "price_range": {
            "type": "string",
            "description": "Price range"
          },
          "available_days": {
            "type": "string",
            "description": "Available days"
          },
          "available_hours": {
            "type": "string",
            "description": "Available hours"
          },
          "rating": {
            "type": "number",
            "description": "Rating (0-5)"
          },
          "total_jobs": {
            "type": "integer",
            "description": "Total completed jobs"
          },
          "is_active": {
            "type": "boolean",
            "description": "Active status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Customer ID"
          },
          "name": {
            "type": "string",
            "description": "Customer name"
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "line_id": {
            "type": "string",
            "description": "LINE ID"
          },
          "location": {
            "type": "string",
            "description": "Address"
          },
          "district": {
            "type": "string",
            "description": "District"
          },
          "subdistrict": {
            "type": "string",
            "description": "Subdistrict"
          },
          "province": {
            "type": "string",
            "description": "Province"
          },
          "service_category_id": {
            "type": "integer",
            "description": "Service category ID"
          },
          "job_description": {
            "type": "string",
            "description": "Job description"
          },
          "budget_range": {
            "type": "string",
            "description": "Budget range"
          },
          "urgency_level": {
            "type": "string",
            "enum": ["low", "medium", "high"],
            "description": "Urgency level"
          },
          "preferred_contact": {
            "type": "string",
            "enum": ["phone", "line", "both"],
            "description": "Preferred contact method"
          },
          "is_active": {
            "type": "boolean",
            "description": "Active status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/providers": {
      "get": {
        "summary": "Get all service providers",
        "tags": ["Providers"],
        "parameters": [
          {
            "in": "query",
            "name": "category_id",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by service category"
          },
          {
            "in": "query",
            "name": "district",
            "schema": {
              "type": "string"
            },
            "description": "Filter by district"
          },
          {
            "in": "query",
            "name": "subdistrict",
            "schema": {
              "type": "string"
            },
            "description": "Filter by subdistrict"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by name or description"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of providers"
          }
        }
      },
      "post": {
        "summary": "Create new provider",
        "tags": ["Providers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Provider"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/customers": {
      "get": {
        "summary": "Get all customers",
        "tags": ["Customers"],
        "parameters": [
          {
            "in": "query",
            "name": "category_id",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by service category"
          },
          {
            "in": "query",
            "name": "district",
            "schema": {
              "type": "string"
            },
            "description": "Filter by district"
          },
          {
            "in": "query",
            "name": "urgency_level",
            "schema": {
              "type": "string",
              "enum": ["low", "medium", "high"]
            },
            "description": "Filter by urgency level"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by name or job description"
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers"
          }
        }
      },
      "post": {
        "summary": "Create new customer",
        "tags": ["Customers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully"
          }
        }
      }
    },
...
    "/api/auto-matches": {
      "get": {
        "summary": "Get auto-generated matches",
        "tags": ["Matches"],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of auto-generated matches"
          }
        }
      }
    },
...
  }
}